Validators were added to the website and foundation fields: A URLValidator was added to the website field to ensure that the website URL is valid. A MaxValueValidator was added to the foundation field to ensure that the date of foundation is not in the future.
An identifier field was added to the model: An identifier field was added to the model to generate a unique identifier for each company. This field is automatically generated when a new company is created.

The save method was overridden: The save method was overridden to generate an identifier if one is not already set for the company.

The created_at field was modified: The created_at field was modified to use the auto_now_add=True parameter to automatically set the creation date of the company when it is first created.